Refactor NUT integration to move coordinator to separate module
Move the coordinator logic into a dedicated `coordinator.py` module,
introducing `NutCoordinator` as a proper `DataUpdateCoordinator` subclass
with `config_entry` typed at the class level. Move `NutConfigEntry` type
alias to `coordinator.py` and update all imports accordingly.
